+++ projects/_08_publication/_01_job_role_estimates/tests/fargate/test_02_clean_pub_data.py
@@ -1,4 +1,4 @@
-from unittest.mock import ANY, Mock, patch
+from unittest.mock import Mock, patch
 
 import projects._08_publication._01_job_role_estimates.fargate._02_clean_pub_data as job